requirements.txt
Flask>=2.0.2
pytest~=8.1.1
requests~=2.31.0
pact-python~=2.1.3